Move SocketHeartbeat to OMGSocketClient
