+++ omisego-sdk/src/main/java/co/omisego/omisego/model/socket/SocketTopic.kt
@@ -1,7 +1,7 @@
 package co.omisego.omisego.model.socket
 
 import co.omisego.omisego.websocket.SocketCustomEventCallback
-import co.omisego.omisego.websocket.channel.interval.SocketHeartbeat
+import co.omisego.omisego.websocket.interval.SocketHeartbeat
 
 /*
  * OmiseGO
